vendor: go-git: import IsDir() check for readReferenceFile()
This fix has been accepted and merged upstream; the gist of the problem is
that readReferenceFile() must check that what it's been requested to read
isn't a directory. read() on a directory fd may work on some platforms (e.g.
FreeBSD), so readReferenceFile() would previously succeed at the earliest
point in a branch name (e.g. "feature" in "feature/myfeature") despite the
fact that this isn't a branch ref but a directory.

Fixes issue #9938

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>